TimerTest script now compiles and runs.
